requirements.txt
paho-mqtt==1.6.1
pymodbus==3.6.8
influxdb-client==1.38.0
websockets==12.0
pymupdf==1.23.8
pinecone-client==5.0.1
openai>=1.14.0
python-dotenv>=1.0.0

